requirements.txt
pydub
pygame
Pillow
json
os
webbrowser
random
